#333 Bilder aus verzeichnis löschen

+++ backend/levt/app/Http/Controllers/_ImageController.php
@@ -72,14 +72,24 @@ class _ImageController extends BaseController
         $requestArray = $request->all();
 
         $image = Image::find($requestArray['imageID']);
+        
+        if($image != null){
+            $imageArray = json_decode($image,true);
+            $src = $imageArray['src'];
+            //$src = substr($src,-66);
+            $src = substr($src,107);
+            print_r($src);
+            Storage::delete($src);
 
-        $image->delete();
+            $image->delete();
 
-        $outputArray = [
-            "deleted" => true
-        ];
+            $outputArray = [
+                "deleted" => true
+            ];
+            return json_encode($outputArray,JSON_PRETTY_PRINT);
+        }
 
-        return json_encode($outputArray,JSON_PRETTY_PRINT);
+        
     }
 
 
@@ -105,8 +115,8 @@ class _ImageController extends BaseController
         //print_r($request->file('picUpload'));
 
         $insertImagesArray = [
-            //sftp://flock-1427@flock-1427.students.fhstp.ac.at/flock-1427.students.fhstp.ac.at/backend/storage/app/images/2019/12/11/17/
-            'src' => "/backend/levt/storage/app/".$request->file('picUpload')->store('images/'.$year.'/'.$month.'/'.$day.'/'.$hour),
+            'src' => "sftp://flock-1427@flock-1427.students.fhstp.ac.at/flock-1427.students.fhstp.ac.at/backend/levt/storage/app/"
+            .$request->file('picUpload')->store('images/'.$year.'/'.$month.'/'.$day.'/'.$hour),
             'coordinateX' => $array['coordinateX'],
             'coordinateY' => $array['coordinateY'],
             'date' => $array['date'],
